fix: refactoring the types
