+++ src/background/services/protocols/utils.ts
@@ -2,7 +2,7 @@ import { Group, BigNumberish } from "@semaphore-protocol/group";
 import { MerkleProof } from "@zk-kit/incremental-merkle-tree";
 import { bigintToHex, hexToBigint } from "bigint-conversion";
 
-import { MerkleProofArtifacts, VerificationKeyT } from "@src/types";
+import { MerkleProofArtifacts, VerificationKey } from "@src/types";
 
 export function deserializeMerkleProof(merkleProof: MerkleProof): MerkleProof {
   return {
@@ -59,8 +59,8 @@ export async function getMerkleProof({
       });
 }
 
-export async function getRlnVerficationKeyJson(rlnVerificationKeyPath: string): Promise<VerificationKeyT> {
-  return fetch(rlnVerificationKeyPath).then((res) => res.json() as Promise<VerificationKeyT>);
+export async function getRlnVerficationKeyJson(rlnVerificationKeyPath: string): Promise<VerificationKey> {
+  return fetch(rlnVerificationKeyPath).then((res) => res.json() as Promise<VerificationKey>);
 }
 
 async function getRemoteMerkleProof(merkleStorageAddress: string, identityCommitmentHex: string): Promise<MerkleProof> {
