+++ src/types/contentScript/index.ts
@@ -1,0 +1,6 @@
+export interface InjectedMessageData<M = unknown, P = unknown> {
+  target: string;
+  nonce: string;
+  message: M;
+  payload: [string, P];
+}
