+++ src/types/proof/Merkle.ts
@@ -1,0 +1,7 @@
+export type { MerkleProof } from "@zk-kit/incremental-merkle-tree";
+
+export type MerkleProofArtifacts = {
+  leaves: string[];
+  depth: number;
+  leavesPerNode: number;
+};
