+++ src/types/proof/RLNProof.ts
@@ -1,0 +1,7 @@
+import { SemaphoreProofRequest } from "./SemaphoreProof";
+
+export type { RLNFullProof, VerificationKey } from "rlnjs";
+
+export interface RLNProofRequest extends SemaphoreProofRequest {
+  rlnIdentifier: string;
+}
