+++ src/types/proof/index.ts
@@ -1,0 +1,4 @@
+export * from "./Merkle";
+export * from "./RLNProof";
+export * from "./SemaphoreProof";
+export * from "./ZkProof";
