+++ src/types/reactComponent/index.ts
@@ -1,0 +1,12 @@
+import { IconProp } from "@fortawesome/fontawesome-svg-core";
+
+export type SelectOption = {
+  readonly value: string;
+  readonly label: string;
+  readonly icon: IconProp | null;
+};
+
+export interface PasswordFormFields {
+  password: string;
+  confirmPassword: string;
+}
