+++ src/types/utility/index.ts
@@ -1,0 +1,5 @@
+export interface DeferredPromise<T> {
+  promise: Promise<T>;
+  resolve?: (arg: T | PromiseLike<T>) => void;
+  reject?: (reason?: unknown) => void;
+}
